@@ -8,8 +8,8 @@
     "clean": "rimraf dist && mkdir dist",
     "prestart": "npm run clean && cp -R app/runtime/* dist && cd dist && npm install",
     "start": "cross-env NODE_ENV=development concurrently --kill-others \"npm run webpack -- --watch\" \"node dist/server.js\"",
-    "eslint": "eslint app/**/**.js --fix",
-    "prewebpack": "npm run eslint",
+    "tslint": "tslint -c tslint.json --fix 'app/**/*.{ts,tsx}'",
+    "prewebpack": "npm run tslint",
     "webpack": "webpack --config ./webpack.config.js --profile --colors",
     "prebuild": "npm run clean",
     "build": "cross-env NODE_ENV=production npm run webpack",
@@ -24,54 +24,49 @@
   "dependencies": {
     "@types/moment-timezone": "0.2.34",
     "@types/object-assign": "4.0.30",
-    "@types/react": "15.0.12",
-    "@types/react-dom": "0.14.23",
-    "@types/react-redux": "4.4.36",
-    "@types/react-router": "3.0.5",
-    "@types/requirejs": "2.1.28",
+    "@types/react": "15.0.38",
+    "@types/react-dom": "15.5.1",
+    "@types/react-redux": "4.4.46",
+    "@types/react-router": "4.0.14",
+    "@types/requirejs": "2.1.29",
     "aws-sdk": "2.4.13",
     "bootstrap": "3.3.7",
     "exceljs": "0.2.24",
-    "font-awesome": "4.6.3",
-    "moment": "2.17.1",
-    "moment-timezone": "0.5.11",
-    "react": "15.2.1",
-    "react-bootstrap": "0.29.5",
-    "react-data-grid": "^1.0.42",
-    "react-date-picker": "5.3.26",
-    "react-dom": "15.2.1",
-    "react-redux": "4.4.5",
-    "react-router": "2.5.1",
-    "redux": "3.5.2",
-    "uuid": "2.0.2"
+    "font-awesome": "4.7.0",
+    "moment": "2.18.1",
+    "moment-timezone": "0.5.13",
+    "react": "15.6.1",
+    "react-bootstrap": "0.31.1",
+    "react-data-grid": "2.0.53",
+    "react-data-grid-addons": "2.0.54",
+    "react-date-picker": "5.3.28",
+    "react-dom": "15.6.1",
+    "react-redux": "5.0.5",
+    "react-router": "4.1.2",
+    "react-router-dom": "4.1.2",
+    "redux": "3.7.2",
+    "uuid": "3.1.0"
   },
   "devDependencies": {
-    "awesome-typescript-loader": "3.0.8",
-    "babel-cli": "6.6.5",
-    "babel-loader": "6.2.4",
-    "babel-plugin-transform-react-display-name": "6.23.0",
-    "babel-preset-es2015": "6.6.0",
-    "babel-preset-react": "6.5.0",
-    "body-parser": "1.15.2",
-    "concurrently": "3.3.0",
-    "copy-webpack-plugin": "3.0.1",
-    "cross-env": "1.0.8",
-    "css-loader": "0.23.1",
-    "eslint": "3.0.1",
-    "eslint-config-google": "0.6.0",
-    "eslint-plugin-react": "5.1.1",
-    "express": "4.13.3",
-    "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.9.0",
-    "html-webpack-plugin": "2.22.0",
-    "json-loader": "0.5.4",
-    "morgan": "1.7.0",
-    "react-transform-catch-errors": "1.0.2",
-    "redux-devtools": "3.3.1",
-    "rimraf": "2.4.3",
-    "source-map-loader": "0.1.6",
-    "style-loader": "0.13.1",
-    "typescript": "2.2.1",
-    "webpack": "1.12.14"
+    "awesome-typescript-loader": "3.2.1",
+    "body-parser": "1.17.2",
+    "concurrently": "3.5.0",
+    "copy-webpack-plugin": "4.0.1",
+    "cross-env": "5.0.1",
+    "css-loader": "0.28.4",
+    "express": "4.15.3",
+    "extract-text-webpack-plugin": "2.1.0",
+    "file-loader": "0.11.2",
+    "html-webpack-plugin": "2.29.0",
+    "json-loader": "0.5.6",
+    "morgan": "1.8.2",
+    "redux-devtools": "3.4.0",
+    "rimraf": "2.6.1",
+    "source-map-loader": "0.2.1",
+    "style-loader": "0.18.2",
+    "tslint": "5.5.0",
+    "tslint-react": "3.0.0",
+    "typescript": "2.4.2",
+    "webpack": "2.4.1"
   }
 }
