@@ -1,11 +1,12 @@
 /**
  * Created by halversondm on 2/22/17.
  */
-"use strict";
-
 import * as React from "react";
 
-class PrivacyPolicy extends React.Component<any, any> {
+export class PrivacyPolicy extends React.Component<any, any> {
+
+    // tslint:disable-next-line
+    private url: string = "http://consumercal.org/california-online-privacy-protection-act-caloppa/#sthash.0FdRbT51.dpuf";
 
     render() {
         return (
@@ -31,9 +32,9 @@ class PrivacyPolicy extends React.Component<any, any> {
                         <div className="col-2">
                             <div className="quick-links text-center">3rd Party Links</div>
                         </div>
-                        <div className="col-2"></div>
+                        <div className="col-2" />
                     </div>
-                    <div style={{clear: "both", height: "10px"}}></div>
+                    <div style={{ clear: "both", height: "10px" }} />
                     <div className="ppConsistencies">
                         <div className="col-2">
                             <div className="col-12 quick-links2 gen-text-center">Google AdSense</div>
@@ -42,8 +43,8 @@ class PrivacyPolicy extends React.Component<any, any> {
                             <div className="col-12 quick-links2 gen-text-center">
                                 Fair Information Practices
                                 <div className="col-8 gen-text-left gen-xs-text-center"
-                                     style={{fontSize: "12px", position: "relative", left: "20px"}}>
-                                    Fair information<br/> Practices
+                                    style={{ fontSize: "12px", position: "relative", left: "20px" }}>
+                                    Fair information<br /> Practices
                                 </div>
                             </div>
                         </div>
@@ -60,10 +61,10 @@ class PrivacyPolicy extends React.Component<any, any> {
                             </div>
                         </div>
                         <div className="col-2">
-                            <div className="quick-links2 gen-text-center">Our Contact Information<br/></div>
+                            <div className="quick-links2 gen-text-center">Our Contact Information<br /></div>
                         </div>
                     </div>
-                    <div style={{clear: "both", height: "10px"}}></div>
+                    <div style={{ clear: "both", height: "10px" }} />
                     <div className="innerText">This privacy policy has been compiled to better serve those who are
                         concerned with how
                         their
@@ -77,43 +78,43 @@ class PrivacyPolicy extends React.Component<any, any> {
                         to
                         get a clear understanding of how we collect, use, protect or otherwise handle your Personally
                         Identifiable
-                        Information in accordance with our website.<br/></div>
-                    <span id="infoCo"/><br/>
+                        Information in accordance with our website.<br /></div>
+                    <span id="infoCo" /><br />
                     <div className="grayText"><strong>What personal information do we collect from the people that visit
                         our blog,
                         website
                         or app?</strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText">When ordering or registering on our site, as appropriate, you may be
                         asked to enter your
                         name, email address or other details to help you with your experience.
                     </div>
-                    <br/>
+                    <br />
                     <div className="grayText"><strong>When do we collect information?</strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText">We collect information from you when you register on our site or enter
                         information on our
                         site.
                     </div>
-                    <br/> <span id="infoUs"/><br/>
+                    <br /> <span id="infoUs" /><br />
                     <div className="grayText"><strong>How do we use your information? </strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText"> We may use the information we collect from you when you register, make a
                         purchase, sign
                         up
                         for our newsletter, respond to a survey or marketing communication, surf the website, or use
                         certain other
                         site
-                        features in the following ways: <br/>We may contact you about the application.<br/></div>
-                    <span id="infoPro"/><br/>
+                        features in the following ways: <br />We may contact you about the application.<br /></div>
+                    <span id="infoPro" /><br />
                     <div className="grayText"><strong>How do we protect your information?</strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText">We do not use vulnerability scanning and/or scanning to PCI standards.
                     </div>
                     <div className="innerText">We only provide articles and information. We never ask for credit card
                         numbers.
                     </div>
-                    <div className="innerText">We do not use Malware Scanning.<br/><br/></div>
+                    <div className="innerText">We do not use Malware Scanning.<br /><br /></div>
                     <div className="innerText">Your personal information is contained behind secured networks and is
                         only accessible by
                         a
@@ -123,67 +124,68 @@ class PrivacyPolicy extends React.Component<any, any> {
                         via Secure
                         Socket Layer (SSL) technology.
                     </div>
-                    <br/>
+                    <br />
                     <div className="innerText">We implement a variety of security measures when a user enters, submits,
                         or accesses
                         their
                         information to maintain the safety of your personal information.
                     </div>
-                    <br/>
+                    <br />
                     <div className="innerText">All transactions are processed through a gateway provider and are not
                         stored or processed
                         on
                         our servers.
                     </div>
-                    <span id="coUs"/><br/>
+                    <span id="coUs" /><br />
                     <div className="grayText"><strong>Do we use 'cookies'?</strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText">We do not use cookies for tracking purposes</div>
-                    <div className="innerText"><br/>You can choose to have your computer warn you each time a cookie is
+                    <div className="innerText"><br />You can choose to have your computer warn you each time a cookie is
                         being sent, or
                         you
                         can choose to turn off all cookies. You do this through your browser settings. Since each
                         browser is
                         a little
                         different, look at your browser's Help Menu to learn the correct way to modify your
-                        cookies.<br/></div>
-                    <br/>
+                        cookies.<br /></div>
+                    <br />
                     <div className="innerText">If you turn cookies off, some features will be disabled. that make your
                         site experience
                         more
                         efficient and may not function properly.
                     </div>
-                    <br/>
+                    <br />
                     <div className="innerText">However, you will still be able to place orders .</div>
-                    <br/><span id="trDi"/><br/>
+                    <br /><span id="trDi" /><br />
                     <div className="grayText"><strong>Third-party disclosure</strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText">We do not sell, trade, or otherwise transfer to outside parties your
                         Personally
                         Identifiable
                         Information.
                     </div>
-                    <span id="trLi"/><br/>
+                    <span id="trLi" /><br />
                     <div className="grayText"><strong>Third-party links</strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText">We do not include or offer third-party products or services on our
                         website.
                     </div>
-                    <span id="gooAd"/><br/>
+                    <span id="gooAd" /><br />
                     <div className="blueText"><strong>Google</strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText">Google's advertising requirements can be summed up by Google's
                         Advertising Principles.
                         They
                         are put in place to provide a positive experience for users.&nbsp;
-                        <a href="https://support.google.com/adwordspolicy/answer/1316548?hl=en">https://support.google.com/adwordspolicy/answer/1316548?hl=en</a>
-                        <br/><br/></div>
+                        <a href="https://support.google.com/adwordspolicy/answer/1316548?hl=en">
+                            https://support.google.com/adwordspolicy/answer/1316548?hl=en</a>
+                        <br /><br /></div>
                     <div className="innerText">We have not enabled Google AdSense on our site but we may do so in the
                         future.
                     </div>
-                    <span id="calOppa"/><br/>
+                    <span id="calOppa" /><br />
                     <div className="blueText"><strong>California Online Privacy Protection Act</strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText">CalOPPA is the first state law in the nation to require commercial
                         websites and online
                         services to post a privacy policy. The law's reach stretches well beyond California to require
@@ -197,40 +199,42 @@ class PrivacyPolicy extends React.Component<any, any> {
                         shared. -
                         See
                         more at:&nbsp;
-                        <a href="http://consumercal.org/california-online-privacy-protection-act-caloppa/#sthash.0FdRbT51.dpuf">http://consumercal.org/california-online-privacy-protection-act-caloppa/#sthash.0FdRbT51.dpuf</a><br/>
+                        <a href={this.url}>{this.url}</a>
+                        <br />
                     </div>
-                    <div className="innerText"><br/><strong>According to CalOPPA, we agree to the
-                        following:</strong><br/></div>
+                    <div className="innerText"><br /><strong>According to CalOPPA, we agree to the
+                        following:</strong><br /></div>
                     <div className="innerText">Users can visit our site anonymously.</div>
                     <div className="innerText">Once this privacy policy is created, we will add a link to it on our home
                         page or as a
-                        minimum, on the first significant page after entering our website.<br/></div>
+                        minimum, on the first significant page after entering our website.<br /></div>
                     <div className="innerText">Our Privacy Policy link includes the word 'Privacy' and can easily be
                         found on the page
                         specified above.
                     </div>
-                    <div className="innerText"><br/>You will be notified of any Privacy Policy changes:</div>
+                    <div className="innerText"><br />You will be notified of any Privacy Policy changes:</div>
                     <div className="innerText">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; <strong>&bull;</strong> On our Privacy
-                        Policy Page<br/>
+                        Policy Page<br />
                     </div>
                     <div className="innerText">Can change your personal information:</div>
                     <div className="innerText">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; <strong>&bull;</strong> By emailing us
                     </div>
-                    <div className="innerText"><br/><strong>How does our site handle Do Not Track signals?</strong><br/>
+                    <div className="innerText"><br /><strong>How does our site handle Do Not Track signals?</strong>
+                    <br />
                     </div>
                     <div className="innerText">We honor Do Not Track signals and Do Not Track, plant cookies, or use
                         advertising when a
                         Do
                         Not Track (DNT) browser mechanism is in place.
                     </div>
-                    <div className="innerText"><br/><strong>Does our site allow third-party behavioral
-                        tracking?</strong><br/></div>
+                    <div className="innerText"><br /><strong>Does our site allow third-party behavioral
+                        tracking?</strong><br /></div>
                     <div className="innerText">It's also important to note that we do not allow third-party behavioral
                         tracking
                     </div>
-                    <span id="coppAct"/><br/>
+                    <span id="coppAct" /><br />
                     <div className="blueText"><strong>COPPA (Children Online Privacy Protection Act)</strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText">When it comes to the collection of personal information from children
                         under the age of 13
                         years old, the Children's Online Privacy Protection Act (COPPA) puts parents in control. The
@@ -239,13 +243,13 @@ class PrivacyPolicy extends React.Component<any, any> {
                         what
                         operators
                         of websites and online services must do to protect children's privacy and safety
-                        online.<br/><br/></div>
+                        online.<br /><br /></div>
                     <div className="innerText">We do not specifically market to children under the age of 13 years
                         old.
                     </div>
-                    <span id="ftcFip"/><br/>
+                    <span id="ftcFip" /><br />
                     <div className="blueText"><strong>Fair Information Practices</strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText">The Fair Information Practices Principles form the backbone of privacy
                         law in the United
                         States and the concepts they include have played a significant role in the development of data
@@ -254,7 +258,7 @@ class PrivacyPolicy extends React.Component<any, any> {
                         around the globe. Understanding the Fair Information Practice Principles and how they should be
                         implemented
                         is
-                        critical to comply with the various privacy laws that protect personal information.<br/><br/>
+                        critical to comply with the various privacy laws that protect personal information.<br /><br />
                     </div>
                     <div className="innerText"><strong>In order to be in line with Fair Information Practices we will
                         take the following
@@ -263,7 +267,7 @@ class PrivacyPolicy extends React.Component<any, any> {
                     <div className="innerText">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; <strong>&bull;</strong> Within 1 business
                         day
                     </div>
-                    <div className="innerText"><br/>We also agree to the Individual Redress Principle which requires
+                    <div className="innerText"><br />We also agree to the Individual Redress Principle which requires
                         that individuals
                         have
                         the right to legally pursue enforceable rights against data collectors and processors who fail
@@ -276,22 +280,22 @@ class PrivacyPolicy extends React.Component<any, any> {
                         non-compliance by
                         data processors.
                     </div>
-                    <span id="canSpam"/><br/>
+                    <span id="canSpam" /><br />
                     <div className="blueText"><strong>CAN SPAM Act</strong></div>
-                    <br/>
+                    <br />
                     <div className="innerText">The CAN-SPAM Act is a law that sets the rules for commercial email,
                         establishes
                         requirements
                         for commercial messages, gives recipients the right to have emails stopped from being sent to
                         them, and
                         spells
-                        out tough penalties for violations.<br/><br/></div>
+                        out tough penalties for violations.<br /><br /></div>
                     <div className="innerText"><strong>We collect your email address in order to:</strong></div>
                     <div className="innerText">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; <strong>&bull;</strong> Send information,
                         respond to
                         inquiries, and/or other requests or questions
                     </div>
-                    <div className="innerText"><br/><strong>To be in accordance with CANSPAM, we agree to the
+                    <div className="innerText"><br /><strong>To be in accordance with CANSPAM, we agree to the
                         following:</strong></div>
                     <div className="innerText">&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; <strong>&bull;</strong> Not use false or
                         misleading
@@ -321,28 +325,26 @@ class PrivacyPolicy extends React.Component<any, any> {
                         using
                         the link at the bottom of each email.
                     </div>
-                    <div className="innerText"><strong><br/>If at any time you would like to unsubscribe from receiving
+                    <div className="innerText"><strong><br />If at any time you would like to unsubscribe from receiving
                         future emails,
                         you
                         can email us at <a href="mailto:toilet.training.tracker@gmail.com">Toilet Training
                             Tracker</a></strong></div>
                     and we will promptly remove you from <strong>ALL</strong> correspondence.
                 </div>
-                <br/><span id="ourCon"/><br/>
+                <br /><span id="ourCon" /><br />
                 <div className="blueText"><strong>Contacting Us</strong></div>
-                <br/>
+                <br />
                 <div className="innerText">If there are any questions regarding this privacy policy, you may contact us
                     using the
-                    information below.<br/><br/></div>
+                    information below.<br /><br /></div>
                 <div className="innerText">https://toilettracker.halversondm.com/</div>
                 <div className="innerText">2314 Shiloh Dr</div>
                 Aurora, IL 60503
                 <div className="innerText">United States</div>
                 <div className="innerText">toilet.training.tracker@gmail.com</div>
-                <div className="innerText"><br/>Last Edited on 2017-02-21</div>
+                <div className="innerText"><br />Last Edited on 2017-02-21</div>
             </div>
         );
     }
 }
-
-export default PrivacyPolicy;
