@@ -1,20 +1,19 @@
 /**
  * Created by halversondm on 9/9/16.
  */
-"use strict";
-
 import * as React from "react";
+import { Redirect } from "react-router-dom";
 
-interface SignupProps {
-    router: any
-}
-
-class Signup extends React.Component<SignupProps, any> {
+export class Signup extends React.Component<any, any> {
 
     constructor(props) {
         super(props);
-        this.state = {name: "", email: "", key: "", keyRetype: ""};
+        this.state = {name: "", email: "", key: "", keyRetype: "", redirect: false};
         this.signup = this.signup.bind(this);
+        this.nameChange = this.nameChange.bind(this);
+        this.emailChange = this.emailChange.bind(this);
+        this.keyChange = this.keyChange.bind(this);
+        this.keyRetypeChange = this.keyRetypeChange.bind(this);
     }
 
     signup(event) {
@@ -22,13 +21,13 @@ class Signup extends React.Component<SignupProps, any> {
         if (this.state.key !== this.state.keyRetype) {
             return;
         }
-        var data = JSON.stringify(this.state);
-        var xhr = new XMLHttpRequest();
+        const data = JSON.stringify(this.state);
+        const xhr = new XMLHttpRequest();
         xhr.open("POST", "/signup");
         xhr.setRequestHeader("Content-Type", "application/json");
         xhr.onload = () => {
             if (xhr.status >= 200 && xhr.status < 400) {
-                this.props.router.replace("/successSignup");
+                this.setState({redirect: true});
             } else {
                 console.log("unsucc ", xhr.responseText);
             }
@@ -39,7 +38,28 @@ class Signup extends React.Component<SignupProps, any> {
         xhr.send(data);
     }
 
+    nameChange(event) {
+        this.setState({name: event.target.value});
+    }
+
+    emailChange(event) {
+        this.setState({email: event.target.value});
+    }
+
+    keyChange(event) {
+        this.setState({key: event.target.value});
+    }
+
+    keyRetypeChange(event) {
+        this.setState({keyRetype: event.target.value});
+    }
+
     render() {
+        if (this.state.redirect) {
+            return (
+                <Redirect to="/successSignup" />
+            );
+        }
         return (
             <form className="form-horizontal">
                 <div className="form-group">
@@ -48,11 +68,7 @@ class Signup extends React.Component<SignupProps, any> {
                     <div className="col-sm-10">
                         <input type="text" className="form-control" id="name"
                                value={this.state.name}
-                               onChange={event => {
-                                   this.setState({
-                                       name: event.target.value
-                                   });
-                               }}/>
+                               onChange={this.nameChange}/>
                     </div>
                 </div>
                 <div className="form-group">
@@ -61,11 +77,7 @@ class Signup extends React.Component<SignupProps, any> {
                     <div className="col-sm-10">
                         <input type="email" className="form-control" id="email"
                                value={this.state.email}
-                               onChange={event => {
-                                   this.setState({
-                                       email: event.target.value
-                                   });
-                               }}/>
+                               onChange={this.emailChange}/>
                     </div>
                 </div>
                 <div className="form-group">
@@ -74,11 +86,7 @@ class Signup extends React.Component<SignupProps, any> {
                     <div className="col-sm-10">
                         <input type="password" className="form-control" id="password"
                                value={this.state.key}
-                               onChange={event => {
-                                   this.setState({
-                                       key: event.target.value
-                                   });
-                               }}/>
+                               onChange={this.keyChange}/>
                     </div>
                 </div>
                 <div className="form-group">
@@ -87,11 +95,7 @@ class Signup extends React.Component<SignupProps, any> {
                     <div className="col-sm-10">
                         <input type="password" className="form-control" id="password"
                                value={this.state.keyRetype}
-                               onChange={event => {
-                                   this.setState({
-                                       keyRetype: event.target.value
-                                   });
-                               }}/>
+                               onChange={this.keyRetypeChange}/>
                     </div>
                 </div>
                 <div className="form-group">
@@ -101,8 +105,7 @@ class Signup extends React.Component<SignupProps, any> {
                         </button>
                     </div>
                 </div>
-            </form>);
+            </form>
+        );
     }
 }
-
-export default Signup;
