@@ -1,17 +1,15 @@
 /**
  * Created by halversondm on 9/10/16.
  */
-"use strict";
-
 import * as React from "react";
-import {Link} from "react-router";
+import {Link} from "react-router-dom";
 
-class SuccessfulSignup extends React.Component<any, any> {
+export class SuccessfulSignup extends React.Component<any, any> {
 
     render() {
         return (
             <div>
-                Your sign-up was successful.
+                Your sign-up was successful.  Please login.
                 <br/>
                 <Link to="login">
                     <button type="button" className="btn btn-primary">Login</button>
@@ -20,5 +18,3 @@ class SuccessfulSignup extends React.Component<any, any> {
         );
     }
 }
-
-export default SuccessfulSignup;
